requirements.txt
# Core framework (pinned for reproducibility)
fastapi==0.121.0
uvicorn==0.38.0
pydantic==2.12.3
pydantic-settings==2.11.0

# Database (pinned for reproducibility)
SQLAlchemy==2.0.44
asyncpg==0.30.0
pgvector==0.4.1
psycopg[binary]==3.2.12
psycopg-pool==3.2.7  # Connection pooling

# Core dependencies (pinned for reproducibility)
numpy==2.3.4
sentence-transformers==5.1.2  # Embeddings and search
APScheduler==3.11.1
croniter==6.0.0

# Document parsing (pinned for reproducibility)
beautifulsoup4==4.14.2
pdfplumber==0.11.7
PyPDF2==3.0.1
python-docx==1.2.0
openpyxl==3.1.5
xlrd==2.0.2
python-magic==0.4.27
chardet==5.2.0

# Payload loaders (pinned for reproducibility)
boto3==1.40.65
google-cloud-storage==2.18.2
google-api-python-client==2.110.0  # Google Drive API
google-auth-httplib2==0.2.0         # Drive API auth
google-auth-oauthlib==1.2.0         # Drive API OAuth
requests==2.32.5

# LLM providers (pinned for reproducibility)
openai==2.7.1  # OpenAI + Groq (OpenAI-compatible)
# litellm  # Uncomment for multi-provider support (Claude, Gemini, etc.)

# Production hardening (pinned for security)
PyJWT[crypto]==2.8.0  # JWT authentication with RS256 support
redis==5.0.1          # Rate limiting backend
alembic==1.13.0       # Database migrations
prometheus-client==0.21.0  # Observability and metrics

# Optional performance
# faiss-cpu
# numba
# mmh3
scikit-learn==1.6.1  # For evaluation metrics
